♻️ refactor : 댓글 시리얼라이저 파일 분리 및 서비스/뷰 실제 DB 연동
